Struct mayda::uniform::Uniform

pub struct Uniform<B> {
    // some fields omitted
}

The type of a uniform encoded integer array. Designed for moderate compression and efficient decoding through the Encode trait, and efficient random access through the Access and AccessInto traits.

Support is provided for arrays with as many as (237 - 27) entries, or about 512 GiB of u32s. If your application requires more than that, you should probably be designing your own data structure anyway.

Examples

use mayda::{Access, Encode, Uniform};

let input: Vec<u32> = vec![1, 4, 2, 8, 5, 7];
let mut bits = Uniform::new();
bits.encode(&input).unwrap();

let length = bits.len();
assert_eq!(length, 6);

let output = bits.decode();
assert_eq!(input, output);

let value = bits.access(4);
assert_eq!(value, 5);

let range = bits.access(1..4);
assert_eq!(range, vec![4, 2, 8]); 

Performance

Decoding does not allocate except for the return value, and decodes around 16 GiB/s of decoded integers on difficult inputs. Encoding allocates O(n) memory (n in the length of the array), and encodes around 4 GiB/s of decoded integers. Run cargo bench --bench uniform for performance numbers on your setup.

The performance (speed and compression) degrades gradually as the number of integers falls below 128.

Safety

As a general rule, you should not decode or access Uniform objects from untrusted sources.

A Uniform object performs unsafe pointer operations during encoding and decoding. Changing the header information with mut_storage can cause data to be written to or read from arbitrary addresses in memory.

That said, the situation is the same for any of the data structures in the standard library (consider the set_len method of a Vec).

Methods

impl<B: Bits> Uniform<B>

fn new() -> Self

Creates an empty Uniform object.

Examples

use mayda::{Encode, Uniform};

let input: Vec<u32> = vec![1, 4, 2, 8, 5, 7];
let mut bits = Uniform::new();
bits.encode(&input).unwrap();

let bytes = std::mem::size_of_val(bits.storage());
assert_eq!(bytes, 16);

fn from_slice(slice: &[B]) -> Result<Self, Error>

Creates a Uniform object that encodes the slice.

Examples

use mayda::{Encode, Uniform};

let input: Vec<u32> = vec![1, 5, 7, 15, 20, 27];
let bits = Uniform::from_slice(&input).unwrap();

let output = bits.decode();
assert_eq!(input, output);

fn is_empty(&self) -> bool

Returns true when there are no encoded entries.

Examples

use mayda::Uniform;

let mut bits = Uniform::<u32>::new();
assert_eq!(bits.is_empty(), true);

fn len(&self) -> usize

Returns the number of encoded entries. Note that since the length has to be calculated, Uniform::len() is more expensive than Slice::len().

Examples

use mayda::{Encode, Uniform};

let input: Vec<u32> = vec![1, 4, 2, 8, 5, 7];
let mut bits = Uniform::new();
bits.encode(&input).unwrap();

assert_eq!(bits.len(), 6);

fn storage(&self) -> &[u32]

Exposes the word storage of the Uniform object.

Examples

use mayda::{Encode, Uniform};

let input: Vec<u32> = vec![1, 4, 2, 8, 5, 7];
let mut bits = Uniform::new();
bits.encode(&input).unwrap();

let storage = bits.storage();
assert_eq!(storage.len(), 4);

unsafe fn mut_storage(&mut self) -> &mut Box<[u32]>

Exposes the mutable word storage of the Uniform object.

Safety

A Uniform object performs unsafe pointer operations during encoding and decoding. Changing the header information can cause data to be written to or read from arbitrary addresses in memory.

fn required_width(&self) -> u32

Returns the width of the encoded integer type.

Examples

use mayda::{Encode, Uniform};

let input: Vec<u32> = vec![1, 4, 2, 8, 5, 7];
let mut bits = Uniform::new();
bits.encode(&input).unwrap();

assert_eq!(bits.required_width(), 32);