Trait mayda::utility::AccessInto

pub trait AccessInto<Idx, B: Bits>: Access<Idx> {
    fn access_into(&self, Idx, &mut [B]) -> usize;
}

A trait for indexing a range of an encoded vector. Writes the result into the slice provided. Similar to but less convenient than Index.

Panics

By convention, panics when the index is out of bounds or the slice is of insufficient length.

The default implementations of access_into panic to indicate that there is no available specialization for the type. This should not happen unless there is a library bug.

Required Methods

fn access_into(&self, Idx, &mut [B]) -> usize

The method for the access foo.access_into(bar, slice) operation.